feat: first version of upper deck
