+++ fbw-a32nx/src/systems/instruments/src/EFB/Ground/Pages/Payload/A380_842/Seating/CargoWidget.tsx
@@ -1,8 +1,7 @@
 /* eslint-disable max-len */
 import React from 'react';
-import { BriefcaseFill, CaretDownFill } from 'react-bootstrap-icons';
 import { CargoStationInfo } from '../../Seating/Constants';
-import { ProgressBar } from '../../../../../UtilComponents/Progress/Progress';
+import { CargoBar } from '../../PayloadElements';
 
 interface SeatMapProps {
     cargo: number[],
@@ -21,60 +20,11 @@ enum CargoStation {
 export const CargoWidget: React.FC<SeatMapProps> = ({ cargo, cargoDesired, cargoMap, onClickCargo }) => (
     <>
         <div className="flex absolute top-4 left-40 flex-row px-4 w-fit">
-            <div><BriefcaseFill size={25} className="my-1 mx-3" /></div>
-            <div className="cursor-pointer" onClick={(e) => onClickCargo(CargoStation.FwdBag, e)}>
-                <ProgressBar
-                    height="20px"
-                    width={`${cargoMap[CargoStation.FwdBag].progressBarWidth}px`}
-                    displayBar={false}
-                    completedBarBegin={100}
-                    isLabelVisible={false}
-                    bgcolor="var(--color-highlight)"
-                    completed={cargo[CargoStation.FwdBag] / cargoMap[CargoStation.FwdBag].weight * 100}
-                />
-                <CaretDownFill
-                    size={25}
-                    className="absolute top-0"
-                    style={{ transform: `translateY(-12px) translateX(${cargoDesired[CargoStation.FwdBag] / cargoMap[CargoStation.FwdBag].weight * cargoMap[CargoStation.FwdBag].progressBarWidth - 12}px)` }}
-                />
-            </div>
+            <CargoBar cargoId={CargoStation.FwdBag} cargo={cargo} cargoDesired={cargoDesired} cargoMap={cargoMap} onClickCargo={onClickCargo} />
         </div>
         <div className="flex absolute top-4 left-2/3 flex-row px-4 w-fit">
-            <div><BriefcaseFill size={25} className="my-1 mx-3" /></div>
-            <div className="flex flex-row mr-3 cursor-pointer" onClick={(e) => onClickCargo(CargoStation.AftBag, e)}>
-                <ProgressBar
-                    height="20px"
-                    width={`${cargoMap[CargoStation.AftBag].progressBarWidth}px`}
-                    displayBar={false}
-                    completedBarBegin={100}
-                    isLabelVisible={false}
-                    bgcolor="var(--color-highlight)"
-                    completed={cargo[CargoStation.AftBag] / cargoMap[CargoStation.AftBag].weight * 100}
-                />
-                <CaretDownFill
-                    size={25}
-                    className="absolute top-0"
-                    style={{ transform: `translateY(-12px) translateX(${cargoDesired[CargoStation.AftBag] / cargoMap[CargoStation.AftBag].weight * cargoMap[CargoStation.AftBag].progressBarWidth - 12}px)` }}
-                />
-            </div>
-            <div className="flex flex-row mr-3 cursor-pointer" onClick={(e) => onClickCargo(CargoStation.AftBulk, e)}>
-                <ProgressBar
-                    height="20px"
-                    width={`${cargoMap[CargoStation.AftBulk].progressBarWidth}px`}
-                    displayBar={false}
-                    completedBarBegin={100}
-                    isLabelVisible={false}
-                    bgcolor="var(--color-highlight)"
-                    completed={cargo[CargoStation.AftBulk] / cargoMap[CargoStation.AftBulk].weight * 100}
-                />
-                <CaretDownFill
-                    size={25}
-                    className="absolute top-0"
-                    style={
-                        { transform: `translateY(-12px) translateX(${cargoDesired[CargoStation.AftBulk] / cargoMap[CargoStation.AftBulk].weight * cargoMap[CargoStation.AftBulk].progressBarWidth - 12}px)` }
-                    }
-                />
-            </div>
+            <CargoBar cargoId={CargoStation.AftBag} cargo={cargo} cargoDesired={cargoDesired} cargoMap={cargoMap} onClickCargo={onClickCargo} />
+            <CargoBar cargoId={CargoStation.AftBulk} cargo={cargo} cargoDesired={cargoDesired} cargoMap={cargoMap} onClickCargo={onClickCargo} />
         </div>
     </>
 );
