+++ fbw-a32nx/src/systems/instruments/src/EFB/Ground/Pages/Payload/Seating/Constants.ts
@@ -1,4 +1,13 @@
-export const TYPE = Object.freeze({ NB_ECO: 0, NB_ECO_EMERG: 1, WB_ECO: 2, WB_ECO_EMERG: 3 });
+export enum TYPE {
+    NB_ECO = 0,
+    NB_ECO_EMERG = 1,
+    WB_ECO = 2,
+    WB_ECO_EMERG = 3,
+    WB_BIZ_FLAT_R = 4,
+    WB_BIZ_FLAT_L = 5,
+    WB_SUITE_R = 6,
+    WB_SUITE_L = 7,
+}
 
 export const CanvasConst = Object.freeze({
     width: 1000,
@@ -63,6 +72,46 @@ export const SeatConstants = Object.freeze({
         padY: 0,
         imageX: 16,
         imageY: 12.125,
+    },
+    [TYPE.WB_ECO_EMERG]: {
+        len: 16,
+        wid: 12.125,
+        padX: 2,
+        padY: 0,
+        imageX: 16,
+        imageY: 12.125,
+    },
+    [TYPE.WB_BIZ_FLAT_R]: {
+        len: 24,
+        wid: 23.22,
+        padX: 12,
+        padY: 0,
+        imageX: 24,
+        imageY: 23.22,
+    },
+    [TYPE.WB_BIZ_FLAT_L]: {
+        len: 24,
+        wid: 23.22,
+        padX: 12,
+        padY: 0,
+        imageX: 24,
+        imageY: 23.22,
+    },
+    [TYPE.WB_SUITE_R]: {
+        len: 35,
+        wid: 20,
+        padX: 2,
+        padY: 0,
+        imageX: 50,
+        imageY: 50,
+    },
+    [TYPE.WB_SUITE_L]: {
+        len: 35,
+        wid: 20,
+        padX: 5,
+        padY: 0,
+        imageX: 50,
+        imageY: 50,
     },
 });
