+++ fbw-a32nx/src/wasm/systems/a320_systems/src/airframe/mod.rs
@@ -222,9 +222,6 @@ impl A320Airframe {
 
         self.set_target_tow(target_tow);
         self.set_target_to_cg_percent_mac(target_to_cg);
-
-        // println!("ZFW CG MAC: {}", self.zfw_cg_mac());
-        // println!("GW CG MAC: {}", self.gw_cg_mac());
     }
 }
 impl SimulationElement for A320Airframe {
